Update cl k8s use case based on policy changes
