+++ helm-override/pythonsdk-tests/onap-override.yaml
@@ -290,6 +290,14 @@ policy:
     enabled: true
   policy-clamp-ac-k8s-ppnt:
     enabled: true
+    image: onap/policy-clamp-ac-k8s-ppnt:latest
+    repoList:
+      helm:
+        repos:
+          - repoName: chartmuseum
+            address: http://localhost:8080
+        protocols:
+          - http
   policy-gui:
     enabled: false
   policy-nexus:
