+++ test/pythonsdk/src/oransdk/policy/templates/commission_k8s.json.j2
@@ -29,7 +29,7 @@
           "provider": "ONAP",
           "chart": {
             "repository": {
-              "address": "{{chartmuseumIp}}",
+              "address": "http://{{chartmuseumIp}}:{{chartmuseumPort}}",
               "repoName": "chartmuseum",
               "port": {{chartmuseumPort}},
               "protocol": "http"
